// datacenter-ui/src/app/modules/core/api/v1/model/portTypeEnum.ts
/**
 * Datacenter API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `RJ45` - RJ45 (1GbE) * `SFP` - SFP (1G) * `SFP+` - SFP+ (10G) * `SFP28` - SFP28 (25G) * `QSFP+` - QSFP+ (40G) * `QSFP28` - QSFP28 (100G) * `QSFP-DD` - QSFP-DD (400G) * `LC` - LC Fiber * `SC` - SC Fiber * `FC` - Fibre Channel * `USB-A` - USB-A * `USB-C` - USB-C * `SERIAL` - Serial Console * `MGMT` - Management * `HDMI` - HDMI * `VGA` - VGA * `OTHER` - Other
 */
export const PortTypeEnum = {
    Rj45: 'RJ45',
    Sfp: 'SFP',
    Sfp2: 'SFP+',
    Sfp28: 'SFP28',
    Qsfp: 'QSFP+',
    Qsfp28: 'QSFP28',
    QsfpDd: 'QSFP-DD',
    Lc: 'LC',
    Sc: 'SC',
    Fc: 'FC',
    UsbA: 'USB-A',
    UsbC: 'USB-C',
    Serial: 'SERIAL',
    Mgmt: 'MGMT',
    Hdmi: 'HDMI',
    Vga: 'VGA',
    Other: 'OTHER'
} as const;
export type PortTypeEnum = typeof PortTypeEnum[keyof typeof PortTypeEnum];